Add logic for the gameplay flow
I relied heavily on the example article for making the gameplay flow
work for at least marking the gameboard by entering a row and column.
This felt wrong, but a quick look up on Discord made me feel like it was
okay to borrow these ideas as the alternative is to come up with a brand
new way of doing this.

Next step is to add win logic.