Add reset functionality to a won or drawn game
Used window.location.reload(); on a button on the winning or drawn modal
as I did not know any other way to reset the full game. The reason I
did this was because just rerunning the screenController function was
causing some strange functionality were the squares weren't fully
cleared as I was getting an alert for choosing the same square when it
was blank.